cleanup, do things in parallel, config for max items fetched
