+++ rust/s3heap/src/lib.rs
@@ -70,7 +70,7 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use backon::{ExponentialBuilder, Retryable};
-use chroma_storage::{Storage, StorageError};
+use chroma_storage::Storage;
 use chrono::{DateTime, Utc};
 use uuid::Uuid;
 
@@ -97,6 +97,9 @@ pub enum Error {
     /// Internal implementation error
     #[error("internal error: {0}")]
     Internal(String),
+    /// Invalid bucket timestamp (not minute-aligned)
+    #[error("invalid bucket: {0}")]
+    InvalidBucket(String),
     /// Storage backend error
     #[error("storage error: {0}")]
     Storage(#[from] chroma_storage::StorageError),
@@ -171,19 +174,22 @@ impl RetryConfig {
 /// let limits = Limits::default();
 ///
 /// // Create custom limits
-/// let custom_limits = Limits::with_buckets(100);
+/// let custom_limits = Limits::default().with_buckets(100);
 /// ```
 #[derive(Clone, Debug, Default, PartialEq, Eq, Hash)]
 pub struct Limits {
     /// Maximum number of buckets to read during a scan operation.
     /// If None, defaults to 1000 buckets.
-    buckets_to_read: Option<usize>,
+    pub buckets_to_read: Option<usize>,
+    /// Maximum number of items to return.
+    /// If None, returns all items found within bucket limits.
+    pub max_items: Option<usize>,
 }
 
 impl Limits {
     const DEFAULT_BUCKETS_TO_READ: usize = 1000;
 
-    /// Create limits with a specific maximum number of buckets to read.
+    /// Set the maximum number of buckets to read.
     ///
     /// # Arguments
     /// * `max_buckets` - Maximum number of buckets to scan
@@ -193,13 +199,29 @@ impl Limits {
     /// ```
     /// use s3heap::Limits;
     ///
-    /// let limits = Limits::with_buckets(50);
+    /// let limits = Limits::default().with_buckets(50);
     /// assert_eq!(limits.max_buckets(), 50);
     /// ```
-    pub fn with_buckets(max_buckets: usize) -> Self {
-        Self {
-            buckets_to_read: Some(max_buckets),
-        }
+    pub fn with_buckets(mut self, max_buckets: usize) -> Self {
+        self.buckets_to_read = Some(max_buckets);
+        self
+    }
+
+    /// Set the maximum number of items to return.
+    ///
+    /// # Arguments
+    /// * `max_items` - Maximum number of items to return
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use s3heap::Limits;
+    ///
+    /// let limits = Limits::default().with_items(100);
+    /// ```
+    pub fn with_items(mut self, max_items: usize) -> Self {
+        self.max_items = Some(max_items);
+        self
     }
 
     /// Get the maximum number of buckets to read.
@@ -214,7 +236,7 @@ impl Limits {
     /// let default_limits = Limits::default();
     /// assert_eq!(default_limits.max_buckets(), 1000);
     ///
-    /// let custom_limits = Limits::with_buckets(50);
+    /// let custom_limits = Limits::default().with_buckets(50);
     /// assert_eq!(custom_limits.max_buckets(), 50);
     /// ```
     pub fn max_buckets(&self) -> usize {
@@ -405,11 +427,7 @@ impl HeapWriter {
     /// );
     /// ```
     pub fn new(prefix: String, storage: Storage, heap_scheduler: Arc<dyn HeapScheduler>) -> Self {
-        assert!(!prefix.is_empty(), "prefix cannot be empty");
-        assert!(
-            !prefix.contains("//"),
-            "prefix cannot contain double slashes"
-        );
+        validate_prefix(&prefix);
         Self {
             internal: Internal::new(prefix, storage, heap_scheduler, RetryConfig::default()),
         }
@@ -476,15 +494,45 @@ impl HeapWriter {
             buckets.entry(bucket).or_default().push(heap_item);
         }
 
-        for (bucket, entries) in buckets {
-            self.internal.merge_on_s3(bucket, &entries).await?;
-        }
+        // Execute bucket writes in parallel with concurrency limit
+        use futures::stream::{self, StreamExt, TryStreamExt};
+
+        stream::iter(buckets.into_iter())
+            .map(|(bucket, entries)| async move {
+                self.internal.merge_on_s3(bucket, &entries).await
+            })
+            .buffer_unordered(10)
+            .try_collect::<Vec<_>>()
+            .await?;
 
         Ok(())
     }
 }
 
 //////////////////////////////////////////// HeapPruner ////////////////////////////////////////////
+
+/// Statistics from a pruning operation.
+#[derive(Debug, Clone, Default)]
+pub struct PruneStats {
+    /// Number of items that were pruned (removed)
+    pub items_pruned: usize,
+    /// Number of items that were retained
+    pub items_retained: usize,
+    /// Number of buckets that were deleted entirely
+    pub buckets_deleted: usize,
+    /// Number of buckets that were updated
+    pub buckets_updated: usize,
+}
+
+impl PruneStats {
+    /// Combine stats from multiple pruning operations.
+    pub fn merge(&mut self, other: &PruneStats) {
+        self.items_pruned += other.items_pruned;
+        self.items_retained += other.items_retained;
+        self.buckets_deleted += other.buckets_deleted;
+        self.buckets_updated += other.buckets_updated;
+    }
+}
 
 /// Manages garbage collection of completed tasks from the heap.
 ///
@@ -517,7 +565,7 @@ impl HeapWriter {
 /// pruner.prune(Limits::default()).await?;
 ///
 /// // Prune with custom limits
-/// pruner.prune(Limits::with_buckets(100)).await?;
+/// pruner.prune(Limits::default().with_buckets(100)).await?;
 /// ```
 pub struct HeapPruner {
     internal: Internal,
@@ -550,11 +598,7 @@ impl HeapPruner {
     /// );
     /// ```
     pub fn new(prefix: String, storage: Storage, heap_scheduler: Arc<dyn HeapScheduler>) -> Self {
-        assert!(!prefix.is_empty(), "prefix cannot be empty");
-        assert!(
-            !prefix.contains("//"),
-            "prefix cannot contain double slashes"
-        );
+        validate_prefix(&prefix);
         Self {
             internal: Internal::new(prefix, storage, heap_scheduler, RetryConfig::default()),
         }
@@ -572,7 +616,7 @@ impl HeapPruner {
     /// # Arguments
     ///
     /// * `limits` - Controls how many buckets to scan. Use [`Limits::default()`]
-    ///   for the default of 1000 buckets, or [`Limits::with_buckets()`] for a custom limit.
+    ///   for the default of 1000 buckets, or use `.with_buckets()` for a custom limit.
     ///
     /// # Errors
     ///
@@ -589,18 +633,16 @@ impl HeapPruner {
     /// pruner.prune(Limits::default()).await?;
     ///
     /// // Prune only the first 50 buckets for faster operation
-    /// pruner.prune(Limits::with_buckets(50)).await?;
+    /// pruner.prune(Limits::default().with_buckets(50)).await?;
     /// ```
-    pub async fn prune(&self, limits: Limits) -> Result<(), Error> {
+    pub async fn prune(&self, limits: Limits) -> Result<PruneStats, Error> {
         let buckets = self.internal.list_approx_first_1k_buckets().await?;
-        for bucket in buckets.into_iter().take(
-            limits
-                .buckets_to_read
-                .unwrap_or(Limits::DEFAULT_BUCKETS_TO_READ),
-        ) {
-            self.prune_bucket(bucket).await?;
+        let mut total_stats = PruneStats::default();
+        for bucket in buckets.into_iter().take(limits.max_buckets()) {
+            let stats = self.prune_bucket(bucket).await?;
+            total_stats.merge(&stats);
         }
-        Ok(())
+        Ok(total_stats)
     }
 
     /// Prune completed tasks from a single bucket.
@@ -611,9 +653,9 @@ impl HeapPruner {
     /// * `bucket` - The bucket timestamp to prune
     ///
     /// # Returns
-    /// * `Ok(())` if the bucket was successfully pruned
+    /// * `Ok(PruneStats)` with statistics about the pruning operation
     /// * `Err` if there was an error during pruning
-    async fn prune_bucket(&self, bucket: DateTime<Utc>) -> Result<(), Error> {
+    pub async fn prune_bucket(&self, bucket: DateTime<Utc>) -> Result<PruneStats, Error> {
         let backoff = RetryConfig::default().to_backoff();
 
         (|| async { self.prune_bucket_inner(bucket).await })
@@ -622,32 +664,47 @@ impl HeapPruner {
     }
 
     /// Internal function that performs a single attempt to prune a bucket.
-    async fn prune_bucket_inner(&self, bucket: DateTime<Utc>) -> Result<(), Error> {
+    async fn prune_bucket_inner(&self, bucket: DateTime<Utc>) -> Result<PruneStats, Error> {
         let heap_scheduler = self.internal.heap_scheduler();
-        let (entries, e_tag) = match self.internal.load_bucket(bucket).await {
-            Ok((entries, e_tag)) => (entries, e_tag),
-            Err(Error::Storage(StorageError::NotFound { .. })) => return Ok(()),
-            Err(err) => {
-                return Err(err);
-            }
-        };
-
-        let mut to_retain = Vec::with_capacity(entries.len());
-        for entry in entries {
-            if !heap_scheduler.is_done(&entry.trigger, entry.nonce).await? {
-                to_retain.push(entry);
-            }
+        let (entries, e_tag) = self.internal.load_bucket_or_empty(bucket).await?;
+        if entries.is_empty() {
+            return Ok(PruneStats::default());
         }
 
-        if to_retain.is_empty() {
+        // Check items in parallel for better performance
+        use futures::stream::{self, StreamExt, TryStreamExt};
+
+        let original_count = entries.len();
+        let to_retain: Vec<HeapItem> = stream::iter(entries.into_iter())
+            .map(|entry| async move {
+                let should_retain = !heap_scheduler.is_done(&entry.trigger, entry.nonce).await?;
+                Ok::<_, Error>((entry, should_retain))
+            })
+            .buffer_unordered(10)
+            .try_filter_map(|(entry, should_retain)| async move {
+                Ok(if should_retain { Some(entry) } else { None })
+            })
+            .try_collect()
+            .await?;
+
+        let (buckets_deleted, buckets_updated) = if to_retain.is_empty() {
             self.internal.clear_bucket(bucket).await?;
+            (1, 0)
         } else {
             self.internal
                 .store_bucket(bucket, &to_retain, e_tag)
                 .await?;
-        }
+            (0, 1)
+        };
 
-        Ok(())
+        let stats = PruneStats {
+            items_retained: to_retain.len(),
+            items_pruned: original_count - to_retain.len(),
+            buckets_deleted,
+            buckets_updated,
+        };
+
+        Ok(stats)
     }
 }
 
@@ -685,7 +742,7 @@ impl HeapPruner {
 /// // Get only tasks with a specific name
 /// let specific_tasks = reader.peek(
 ///     |task| task.name == "process_payment",
-///     Limits::with_buckets(100),
+///     Limits::default().with_buckets(100),
 /// ).await?;
 /// ```
 pub struct HeapReader {
@@ -719,11 +776,7 @@ impl HeapReader {
     /// );
     /// ```
     pub fn new(prefix: String, storage: Storage, heap_scheduler: Arc<dyn HeapScheduler>) -> Self {
-        assert!(!prefix.is_empty(), "prefix cannot be empty");
-        assert!(
-            !prefix.contains("//"),
-            "prefix cannot contain double slashes"
-        );
+        validate_prefix(&prefix);
         Self {
             internal: Internal::new(prefix, storage, heap_scheduler, RetryConfig::default()),
         }
@@ -745,7 +798,7 @@ impl HeapReader {
     ///   that should be included in the results. This function is called for
     ///   each non-completed task found.
     /// * `limits` - Controls how many buckets to scan. Use [`Limits::default()`]
-    ///   for the default of 1000 buckets, or [`Limits::with_buckets()`] for a custom limit.
+    ///   for the default of 1000 buckets, or use `.with_buckets()` for a custom limit.
     ///
     /// # Errors
     ///
@@ -764,7 +817,7 @@ impl HeapReader {
     /// // Get only high-priority tasks
     /// let high_priority = reader.peek(
     ///     |task| task.name.starts_with("urgent_"),
-    ///     Limits::with_buckets(50),
+    ///     Limits::default().with_buckets(50),
     /// ).await?;
     ///
     /// // Get tasks for a specific entity
@@ -782,484 +835,38 @@ impl HeapReader {
         let heap_scheduler = self.internal.heap_scheduler();
         let buckets = self.internal.list_approx_first_1k_buckets().await?;
         let mut returns = vec![];
-        for bucket in buckets.into_iter().take(
-            limits
-                .buckets_to_read
-                .unwrap_or(Limits::DEFAULT_BUCKETS_TO_READ),
-        ) {
-            let (entries, _) = match self.internal.load_bucket(bucket).await {
-                Ok((entries, e_tag)) => (entries, e_tag),
-                Err(Error::Storage(StorageError::NotFound { .. })) => continue,
-                Err(err) => {
-                    return Err(err);
-                }
-            };
+        let max_items = limits.max_items.unwrap_or(usize::MAX);
+
+        'outer: for bucket in buckets.into_iter().take(limits.max_buckets()) {
+            let (entries, _) = self.internal.load_bucket_or_empty(bucket).await?;
             for entry in entries {
                 if !heap_scheduler.is_done(&entry.trigger, entry.nonce).await?
                     && should_return(&entry.trigger)
                 {
                     returns.push(entry);
+
+                    // Early termination when we've collected enough items
+                    if returns.len() >= max_items {
+                        break 'outer;
+                    }
                 }
             }
         }
+
         Ok(returns)
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::collections::HashMap;
-    use std::str::FromStr;
-    use std::sync::{Arc, Mutex};
-
-    struct TestScheduler;
-
-    #[async_trait::async_trait]
-    impl HeapScheduler for TestScheduler {
-        async fn is_done(&self, _item: &Triggerable, _nonce: Uuid) -> Result<bool, Error> {
-            Ok(false)
-        }
-
-        async fn next_time_and_nonce(
-            &self,
-            _item: &Triggerable,
-        ) -> Result<Option<(DateTime<Utc>, Uuid)>, Error> {
-            Ok(None)
-        }
-    }
-
-    // More sophisticated test scheduler for comprehensive testing
-    type ScheduleInfo = Option<(DateTime<Utc>, Uuid)>;
-
-    struct ConfigurableScheduler {
-        done_items: Arc<Mutex<HashMap<(Uuid, Uuid), bool>>>,
-        scheduled_items: Arc<Mutex<HashMap<Uuid, ScheduleInfo>>>,
-        error_on_done: Arc<Mutex<bool>>,
-        error_on_schedule: Arc<Mutex<bool>>,
-    }
-
-    impl ConfigurableScheduler {
-        fn new() -> Self {
-            Self {
-                done_items: Arc::new(Mutex::new(HashMap::new())),
-                scheduled_items: Arc::new(Mutex::new(HashMap::new())),
-                error_on_done: Arc::new(Mutex::new(false)),
-                error_on_schedule: Arc::new(Mutex::new(false)),
-            }
-        }
-
-        fn set_error_on_schedule(&self, should_error: bool) {
-            *self.error_on_schedule.lock().unwrap() = should_error;
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl HeapScheduler for ConfigurableScheduler {
-        async fn is_done(&self, item: &Triggerable, nonce: Uuid) -> Result<bool, Error> {
-            if *self.error_on_done.lock().unwrap() {
-                return Err(Error::Internal("Simulated error in is_done".to_string()));
-            }
-            Ok(self
-                .done_items
-                .lock()
-                .unwrap()
-                .get(&(item.uuid, nonce))
-                .copied()
-                .unwrap_or(false))
-        }
-
-        async fn next_time_and_nonce(&self, item: &Triggerable) -> Result<ScheduleInfo, Error> {
-            if *self.error_on_schedule.lock().unwrap() {
-                return Err(Error::Internal(
-                    "Simulated error in next_time_and_nonce".to_string(),
-                ));
-            }
-            Ok(self
-                .scheduled_items
-                .lock()
-                .unwrap()
-                .get(&item.uuid)
-                .cloned()
-                .flatten())
-        }
-    }
-
-    // Tests for prefix validation
-    #[test]
-    #[should_panic(expected = "prefix cannot be empty")]
-    fn writer_panics_on_empty_prefix() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        HeapWriter::new(String::new(), storage, scheduler);
-    }
-
-    #[test]
-    #[should_panic(expected = "prefix cannot be empty")]
-    fn pruner_panics_on_empty_prefix() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        HeapPruner::new(String::new(), storage, scheduler);
-    }
-
-    #[test]
-    #[should_panic(expected = "prefix cannot be empty")]
-    fn reader_panics_on_empty_prefix() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        HeapReader::new(String::new(), storage, scheduler);
-    }
-
-    #[test]
-    fn heap_components_accept_valid_prefix() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-
-        // These should not panic
-        let _writer = HeapWriter::new(
-            "valid-prefix".to_string(),
-            storage.clone(),
-            scheduler.clone(),
-        );
-        let _pruner = HeapPruner::new(
-            "valid-prefix".to_string(),
-            storage.clone(),
-            scheduler.clone(),
-        );
-        let _reader = HeapReader::new("valid-prefix".to_string(), storage, scheduler);
-    }
-
-    #[test]
-    #[should_panic(expected = "prefix cannot contain double slashes")]
-    fn heap_writer_panics_on_double_slash() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        HeapWriter::new("prefix//with//slashes".to_string(), storage, scheduler);
-    }
-
-    #[test]
-    #[should_panic(expected = "prefix cannot contain double slashes")]
-    fn heap_pruner_panics_on_double_slash() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        HeapPruner::new("prefix//with//slashes".to_string(), storage, scheduler);
-    }
-
-    #[test]
-    #[should_panic(expected = "prefix cannot contain double slashes")]
-    fn heap_reader_panics_on_double_slash() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        HeapReader::new("prefix//with//slashes".to_string(), storage, scheduler);
-    }
-
-    // Tests for RetryConfig
-    #[test]
-    fn retry_config_default_values() {
-        let config = RetryConfig::default();
-        assert_eq!(config.min_delay, Duration::from_millis(100));
-        assert_eq!(config.max_delay, Duration::from_secs(10));
-        assert_eq!(config.factor, 2.0);
-        assert_eq!(config.max_retries, 10);
-    }
-
-    #[test]
-    fn retry_config_to_backoff() {
-        let config = RetryConfig {
-            min_delay: Duration::from_millis(50),
-            max_delay: Duration::from_secs(5),
-            factor: 1.5,
-            max_retries: 5,
-        };
-        let backoff = config.to_backoff();
-        // The backoff builder is properly configured
-        // We can't easily test the internal state, but we can verify it builds without panic
-        let _ = backoff;
-    }
-
-    // Tests for Limits
-    #[test]
-    fn limits_default_is_none() {
-        let limits = Limits::default();
-        assert_eq!(limits.buckets_to_read, None);
-    }
-
-    #[test]
-    fn limits_equality() {
-        let limits1 = Limits {
-            buckets_to_read: Some(100),
-        };
-        let limits2 = Limits {
-            buckets_to_read: Some(100),
-        };
-        let limits3 = Limits {
-            buckets_to_read: Some(200),
-        };
-        let limits4 = Limits {
-            buckets_to_read: None,
-        };
-
-        assert_eq!(limits1, limits2);
-        assert_ne!(limits1, limits3);
-        assert_ne!(limits1, limits4);
-    }
-
-    #[test]
-    fn limits_clone() {
-        let original = Limits {
-            buckets_to_read: Some(500),
-        };
-        let cloned = original.clone();
-        assert_eq!(original, cloned);
-        assert_eq!(cloned.buckets_to_read, Some(500));
-    }
-
-    // Tests for Triggerable
-    #[test]
-    fn triggerable_creation_and_equality() {
-        let uuid = Uuid::new_v4();
-        let t1 = Triggerable {
-            uuid,
-            name: "test-task".to_string(),
-        };
-        let t2 = Triggerable {
-            uuid,
-            name: "test-task".to_string(),
-        };
-        let t3 = Triggerable {
-            uuid: Uuid::new_v4(),
-            name: "test-task".to_string(),
-        };
-        let t4 = Triggerable {
-            uuid,
-            name: "different-task".to_string(),
-        };
-
-        assert_eq!(t1, t2);
-        assert_ne!(t1, t3);
-        assert_ne!(t1, t4);
-    }
-
-    #[test]
-    fn triggerable_clone() {
-        let original = Triggerable {
-            uuid: Uuid::new_v4(),
-            name: "clone-test".to_string(),
-        };
-        let cloned = original.clone();
-        assert_eq!(original, cloned);
-        assert_eq!(original.uuid, cloned.uuid);
-        assert_eq!(original.name, cloned.name);
-    }
-
-    #[test]
-    fn triggerable_default() {
-        let t = Triggerable::default();
-        assert_eq!(t.uuid, Uuid::nil());
-        assert_eq!(t.name, "");
-    }
-
-    // Tests for Error enum
-    #[test]
-    fn error_display() {
-        // Test various error types display correctly
-        let etag_error = Error::ETagConflict;
-        assert_eq!(format!("{}", etag_error), "e_tag conflict");
-
-        let missing_etag = Error::MissingETag("test-path".to_string());
-        assert_eq!(format!("{}", missing_etag), "missing e_tag: test-path");
-
-        let internal_error = Error::Internal("something went wrong".to_string());
-        assert_eq!(
-            format!("{}", internal_error),
-            "internal error: something went wrong"
-        );
-
-        let arrow_error = Error::Arrow("column missing".to_string());
-        assert_eq!(format!("{}", arrow_error), "arrow error: column missing");
-
-        // We can't easily construct a uuid::Error, so just test the formatting with a real parse error
-        let uuid_result = Uuid::from_str("invalid-uuid");
-        if let Err(uuid_err) = uuid_result {
-            let error = Error::Uuid(uuid_err);
-            assert!(format!("{}", error).contains("uuid error"));
-        }
-    }
-
-    // Async tests for HeapWriter
-    #[tokio::test]
-    async fn writer_push_empty_items() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        let writer = HeapWriter::new("test-prefix".to_string(), storage, scheduler);
-
-        // Pushing empty items should succeed without doing anything
-        let result = writer.push(&[]).await;
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn writer_push_with_no_scheduled_items() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(ConfigurableScheduler::new());
-        let writer = HeapWriter::new("test-no-schedule".to_string(), storage, scheduler);
-
-        let item = Triggerable {
-            uuid: Uuid::new_v4(),
-            name: "unscheduled".to_string(),
-        };
-
-        // Item has no schedule, so push should succeed but not create any buckets
-        let result = writer.push(&[item]).await;
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn writer_push_with_scheduler_error() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(ConfigurableScheduler::new());
-        scheduler.set_error_on_schedule(true);
-
-        let writer = HeapWriter::new("test-error".to_string(), storage, scheduler);
-
-        let item = Triggerable {
-            uuid: Uuid::new_v4(),
-            name: "error-item".to_string(),
-        };
-
-        // Should propagate the scheduler error
-        let result = writer.push(&[item]).await;
-        assert!(result.is_err());
-        match result {
-            Err(Error::Internal(msg)) => assert!(msg.contains("Simulated error")),
-            _ => panic!("Expected Internal error"),
-        }
-    }
-
-    // Async tests for HeapPruner
-    #[tokio::test]
-    async fn pruner_with_empty_heap() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        let pruner = HeapPruner::new("empty-heap".to_string(), storage, scheduler);
-
-        // Pruning empty heap should succeed
-        let result = pruner.prune(Limits::default()).await;
-        assert!(result.is_ok());
-    }
-
-    #[tokio::test]
-    async fn pruner_respects_limits() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        let pruner = HeapPruner::new("limited-prune".to_string(), storage, scheduler);
-
-        let limits = Limits {
-            buckets_to_read: Some(5),
-        };
-
-        // Should respect the limit even if more buckets exist
-        let result = pruner.prune(limits).await;
-        assert!(result.is_ok());
-    }
-
-    // Async tests for HeapReader
-    #[tokio::test]
-    async fn reader_peek_empty_heap() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        let reader = HeapReader::new("empty-reader".to_string(), storage, scheduler);
-
-        let items = reader.peek(|_| true, Limits::default()).await;
-        assert!(items.is_ok());
-        assert_eq!(items.unwrap().len(), 0);
-    }
-
-    #[tokio::test]
-    async fn reader_peek_with_filter() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        let reader = HeapReader::new("filtered-reader".to_string(), storage, scheduler);
-
-        // Filter that rejects everything
-        let items = reader.peek(|_| false, Limits::default()).await;
-        assert!(items.is_ok());
-        assert_eq!(items.unwrap().len(), 0);
-    }
-
-    #[tokio::test]
-    async fn reader_respects_limits() {
-        let (_temp_dir, storage) = chroma_storage::test_storage();
-        let scheduler = Arc::new(TestScheduler);
-        let reader = HeapReader::new("limited-reader".to_string(), storage, scheduler);
-
-        let limits = Limits {
-            buckets_to_read: Some(3),
-        };
-
-        // Should respect the bucket limit
-        let items = reader.peek(|_| true, limits).await;
-        assert!(items.is_ok());
-    }
-
-    // Edge case tests
-    #[test]
-    fn triggerable_with_empty_name() {
-        let t = Triggerable {
-            uuid: Uuid::new_v4(),
-            name: String::new(),
-        };
-        assert_eq!(t.name, "");
-
-        let t2 = t.clone();
-        assert_eq!(t, t2);
-    }
-
-    #[test]
-    fn triggerable_with_very_long_name() {
-        let long_name = "a".repeat(10000);
-        let t = Triggerable {
-            uuid: Uuid::new_v4(),
-            name: long_name.clone(),
-        };
-        assert_eq!(t.name.len(), 10000);
-        assert_eq!(t.name, long_name);
-    }
-
-    #[test]
-    fn retry_config_with_extreme_values() {
-        let config = RetryConfig {
-            min_delay: Duration::from_nanos(1),
-            max_delay: Duration::from_secs(3600),
-            factor: 100.0,
-            max_retries: 1000,
-        };
-        assert_eq!(config.min_delay, Duration::from_nanos(1));
-        assert_eq!(config.max_delay, Duration::from_secs(3600));
-        assert_eq!(config.factor, 100.0);
-        assert_eq!(config.max_retries, 1000);
-
-        // Should still create a valid backoff
-        let _backoff = config.to_backoff();
-    }
-
-    #[test]
-    fn retry_config_with_zero_retries() {
-        let config = RetryConfig {
-            min_delay: Duration::from_millis(100),
-            max_delay: Duration::from_secs(1),
-            factor: 2.0,
-            max_retries: 0,
-        };
-        assert_eq!(config.max_retries, 0);
-        let _backoff = config.to_backoff();
-    }
-
-    #[test]
-    fn limits_with_max_value() {
-        let limits = Limits {
-            buckets_to_read: Some(usize::MAX),
-        };
-        assert_eq!(limits.buckets_to_read, Some(usize::MAX));
-    }
+/// Validate that a prefix meets the requirements for heap operations.
+///
+/// # Panics
+///
+/// - Panics if `prefix` is empty
+/// - Panics if `prefix` contains "//" (double slashes)
+fn validate_prefix(prefix: &str) {
+    assert!(!prefix.is_empty(), "prefix cannot be empty");
+    assert!(
+        !prefix.contains("//"),
+        "prefix cannot contain double slashes"
+    );
 }
